Fix: #85[유영] 테이블 변경에 따른 방 CRUD 수정

See merge request s09-webmobile1-sub2/S09P12A602!6

+++ backend/oho/src/main/java/com/ssafy/oho/model/dto/request/PlayerRequestDto.java
@@ -11,10 +11,10 @@ import lombok.NoArgsConstructor;
 // @Setter
 @Getter
 public class PlayerRequestDto {
-    private int id;
+    private long id;
     private String nickname;
     private String memberId;
-    private int roomId;
+    private long roomId;
     private boolean head;
     private boolean ready;
     private int score;
